+++ src/collective/liveblog/browser/recent_updates.py
@@ -5,7 +5,6 @@ from collective.liveblog.interfaces import IBrowserLayer
 from collective.liveblog.interfaces import ILiveblog
 from datetime import datetime
 from five import grok
-from plone.memoize import ram
 from time import time
 
 import logging
@@ -13,11 +12,6 @@ import logging
 logger = logging.getLogger(PROJECTNAME)
 
 grok.templatedir('templates')
-
-
-def _updates_since_timestamp_cachekey(method, self, timestamp):
-    return (
-        self.context.absolute_url_path(), int(self.context.modified()), timestamp)
 
 
 class RecentUpdates(grok.View):
@@ -94,21 +88,8 @@ class RecentUpdates(grok.View):
         if self._if_modified_since_request_handler():
             return ''
 
-    def _updates_since_timestamp(self, timestamp):
-        """Return the list of micro-updates since the specified timestamp."""
-        # timestamp should a string representing a float
-        try:
-            float(timestamp)
-        except (TypeError, ValueError):
-            return []
-
-        updates = self.context.restrictedTraverse('view').updates()
-        updates = [u for u in updates if u['timestamp'] > timestamp]
+    def get_latest_microupdates(self):
+        """Return micro-updates posted in the last minute."""
+        updates = self.context.get_microupdates()
+        updates = [u for u in updates if u['timestamp'] > str(time() - 60)]
         return updates
-
-    @ram.cache(_updates_since_timestamp_cachekey)
-    def updates_since_timestamp(self, timestamp):
-        """Return the list of micro-updates since the specified timestamp;
-        the list is cached until a new update is published.
-        """
-        return self._updates_since_timestamp(timestamp)